\name{CropNA}
\alias{CropNA}
\title{Crop outer NA values from a raster}
\description{This function cuts NA values around an 'island' of real values in a Raster* object.}
\usage{CropNA(r, ...)}
\arguments{
  \item{r}{Raster* object}
  \item{\dots}{other arguments, see writeRaster.}
}

\value{a Raster* object with smaller extent.}

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





