\name{greenbrown-package}
\alias{greenbrown-package}
\docType{package}
\title{Trend analysis on earth observation data}
\description{greenbrown is a collection of functions to analyse trends and trend changes in gridded time series like from satellite observations or climate model simulations. 'Greening' describes in the earth observation community positive trends in vegetation greenness. Vegetation greenness can be expressed as Normalized Difference Vegetation Index (NDVI) that is related to the coverage of green vegetation, photosynthetic activity and green biomass. Greening trends were observed worldwide based on satellite observations of the last 30 years (1980-2012). On the other hand, some regions have negative trends in vegetation greenness ('browning'). The package 'greenbrown' provides access to different methods to analyze such greening and browning trends and trend changes in raster datasets like from satellite observations.}

\author{Matthias Forkel <mforkel@bgc-jena.mpg.de>}

\keyword{ package }


