\name{NamesTrendRaster}
\alias{NamesTrendRaster}
\title{Get the layer names for a TrendRaster raster brick}
\description{This function returns the layer names of a raster brick that was created using \code{\link{TrendRaster}}}
\usage{NamesTrendRaster(x)}
\arguments{
  \item{x}{\code{RasterBrick} as created with \code{\link{TrendRaster}} or \code{integer} as the maximum number of breakpoints that was used when the function \code{\link{TrendRaster}} was called.}
}



\author{Matthias Forkel <matthias.forkel@tu-dresden.de> [aut, cre]}



\seealso{\code{\link{TrendRaster}}}

