\name{NoBP}
\alias{NoBP}
\title{Initialize an empty list with breakpoints}
\description{This is an internal function to make an empty list of breakpoints. For the user there is usually no need to use this function.}
\usage{NoBP()}

\value{empty list with slot 'breakpoints'}

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





