\name{NoTrend}
\alias{NoTrend}
\title{Initialize an empty list of trends}
\description{This is an internal function to make an empty list of breakpoints. For the user there is usually no need to use this function.}
\usage{NoTrend(Yt)}
\arguments{
  \item{Yt}{univariate time series of class \code{\link{ts}}}
}

\value{empty list with slot 'breakpoints'}

\author{Matthias Forkel <mforkel@bgc-jena.mpg.de>}





