\name{NoUnc}
\alias{NoUnc}
\title{Initialize an empty data.frame for trend uncertainties}
\description{This is an internal function. For the user there is usually no need to use this function.}
\usage{NoUnc(Yt)}
\arguments{
  \item{Yt}{univariate time series of class \code{\link{ts}}}
}

\value{The function returns a data.frame. }

\author{Matthias Forkel <matthias.forkel@tu-dresden.de> [aut, cre]}





