\name{greenbrown-package}
\alias{greenbrown-package}
\alias{greenbrown}
\docType{package}
\title{Land Surface Phenology and Trend Analysis}
}
\description{Collection of functions to analyse trends, trend changes and phenology events in gridded time series like from satellite observations or climate model simulations. The package provides access to different methods for 1) trend and breakpoint analysis, 2) time series smoothing and interpolation, and 3) analysis of land surface phenology.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{greenbrown}
\packageIndices{greenbrown}
}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}

Maintainer: \packageMaintainer{greenbrown}
}

\keyword{ package }


