\name{greenbrown-package}
\alias{greenbrown-package}
\docType{package}
\title{greenbrown - Land surface phenology and trend analysis}
\description{greenbrown is a collection of functions to analyse trends, trend changes and phenology in gridded time series like from satellite observations or climate model simulations. The package 'greenbrown' provides access to different methods for 1) trend and breakpoint analysis, 2) time series smoothing and interpolation and 3) analysis of phenology metrics.}

\author{Matthias Forkel <mforkel@bgc-jena.mpg.de>}

\keyword{ package }


