\name{NoTrend}
\alias{NoTrend}
\title{Initialize an empty object of class "Trend"}
\description{This is an internal function to make an empty list of class Trend. For the user there is usually no need to use this function.}
\usage{NoTrend(Yt)}
\arguments{
  \item{Yt}{univariate time series of class \code{\link{ts}}}
}

\value{The function returns a list of class "Trend". }

\author{Matthias Forkel <mforkel@bgc-jena.mpg.de>}





