\name{print.Phenology}
\alias{print.Phenology}
\title{Prints phenology metrics}
\description{The function prints an object of class \code{\link{Phenology}}.}
\usage{\method{print}{Phenology}(x, ...)}
\arguments{
  \item{x}{Object of class 'Phenology' as returned from function \code{\link{Phenology}}}
  \item{\dots}{further arguments (not used)}
}



\author{Matthias Forkel <matthias.forkel@tu-dresden.de> [aut, cre]}





